requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --index-url=https://pypi.python.org/simple/ --no-emit-trusted-host --no-index requirements.in
#
asgiref==3.3.1            # via django
django-cors-headers==3.2.1  # via -r requirements.in
django==3.1.4             # via -r requirements.in, django-cors-headers, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
ebooklib==0.17.1          # via -r requirements.in
lxml==4.6.2               # via ebooklib
pytz==2019.3              # via -r requirements.in, django
six==1.14.0               # via -r requirements.in, ebooklib
sqlparse==0.3.0           # via django
tqdm==4.54.1              # via -r requirements.in
